Add files via upload
Data injected from an automated bot responses from Whatsapp. The data is cleaned, and data types are properly assigned for automated marketing blasts. 

Names are cleaned of emojis and a generic substitute is issued for when name is all emojis. name is extracted from emails if email is in place of names.
Numbers are put in a better format with country codes, spaces, dashes, slashes and other symbols are removed.